+++ apiserver/facades/controller/caasprovisioner/provisioner_test.go
@@ -9,7 +9,7 @@ import (
 	"gopkg.in/juju/names.v2"
 
 	"github.com/juju/juju/apiserver/common"
-	"github.com/juju/juju/apiserver/facades/agent/caasprovisioner"
+	"github.com/juju/juju/apiserver/facades/controller/caasprovisioner"
 	"github.com/juju/juju/apiserver/params"
 	apiservertesting "github.com/juju/juju/apiserver/testing"
 	"github.com/juju/juju/state"
